Upstream merge fixes.
If a repo mentioned in an iGOS package.toml does not exist
as github.com/psleng/ just skip it with a warning rather than failing.
Most likely it means that repo needs to be forked.

Also add a warning to manually check that our package.toml files
are up to date with the VyOS equivalents.