live-config build-time emulation.
Instead of passing in "boot=live" to the kernel command line
which effectively makes /lib/live/config/* scripts run via
the live-config.service , do it at buildtime via chroot to
avoid slowing down the boot.

Beware that /lib/live/boot/* scripts, which are supposed to run early
and really do need a running system, are still not handled so we still          need updates/perle-init.sh .  It just does not need to fake up the
/lib/live/config/* actions.

Note that VyOS itself relies on seeing the pattern "BOOT_IMAGE=/boot/"
at the start of /proc/cmdline in order to adjust itself for "image install"
behaviour (which is ultimately also what we want).  However, their code
assumes the boot was via grub, so if it was not (normal for our arm64 boards)
it breaks bootup.  This will have to be readdressed later.