Repo maintenance improvements.
Modify "gitserv" to use a copy of the generated "work" directory.

This is so you can do test builds by pushing temporary changes
to the local repository without polluting the merged work directory.

Specifically, you might need to hack libvyosconfig/build.sh
to include the local test repo versions of vyconf and vyos1x-config
for the build.  Since the build system does the pulling when it
runs, this cannot be done beforehand.