ssh support for rundocker.sh
This will supply the ~/.ssh credentials of the user invoking
rundocker.sh to the container process.

This is so that we can start cloning https://github.com/Perle-Systems-Limited/
repositories in the build.  This is useful for proprietary code which
we do not want public via https://github.com/psleng .

NOTE: Needs commit 0f083a05538f820487670a12375cd57ac68fdf52
from vyos-build to work.  The docker image needs to be rebuilt
with that commit checked out.