Introduce psl-progs to build. (#14)
IMPORTANT: This makes use of the private Perle repository

    git@github.com:Perle-Systems-Limited/

so you must build using an account that has permission
to access that.  The daily build system is being modified
to deal with this.

psl-progs currently includes just a watchdog, which will run as a
system service (perle-watchdog.service).  However, it is structured
so that it is quite easy to add a new subdirectory to build other
PSL proprietary programs which will be added to the resulting .deb

You need only copy/modify the watchdog/Makefile to your new directory
(it needs "all", "install" and "clean" rules) and then the build-deb
script to update postinstall and preremove actions as needed.