Update for merge from upstream.
This brings us up to date as of 20250919.  A full clean build is required.
The following repositories have changed:

        nexus-build
        vyos-1x
        vyos1x-config
        vyconf
        vyatta-cfg
        vyos-build

A clean rebuild is required and as usual if you are in a branch of any of the above you need to merge in changes from psl-master .

There are no CLI changes but many VyOS bugfixes and structural changes:

- libvyosconfig is no longer a submodule of vyos-1x.
  The VyOS folks moved it into vyos-1x directly.
  Our clone of libvyosconfig remains in package-build-iGOS/libvyosconfig
  but is no longer being built.

- Building arm64 via qemu on amd64 is NO LONGER SUPPORTED.
  Various hacks that had been in place for this were causing build issues.
  In particular, we were stuck at an older Go version than VyOS.

- ssh-dss is being deprecated (warning only for now).